pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game"
version = "0.1.0"
description = "Stage 1 implementation for 剑之初程：淬炼之路"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "argon2-cffi>=23.1.0",
  "itsdangerous>=2.2.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[tool.setuptools]
packages = ["engine", "server", "server.app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
